+++ CHANGELOG.md
@@ -22,7 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export to DQX: Correct DQX format for global-level quality check of data contract export. (#877)
 - Import the table tags from a open data contract spec v3 (#895)
 - dbt export: Enhanced model-level primaryKey support with automatic test generation for single and multiple column primary keys (#898)
-
+- ODCS: field discarded when no logicalType defined  (#891)
+ 
 ### Removed
 
 - Removed specific linters, as the linters did not support ODCS (#913)